Move ParsingHook to infra-parser module
