+++ shardingsphere-infra/shardingsphere-infra-parser/src/main/java/org/apache/shardingsphere/infra/parser/hook/ParsingHook.java
@@ -15,62 +15,33 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.hook;
+package org.apache.shardingsphere.infra.parser.hook;
 
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 
-import java.util.Collection;
-import java.util.LinkedList;
-import java.util.ServiceLoader;
-
 /**
- * Parsing hook registry.
+ * Parsing hook.
  */
-public final class ParsingHookRegistry {
-    
-    private static final ParsingHookRegistry INSTANCE = new ParsingHookRegistry();
-    
-    private final Collection<ParsingHook> hooks = new LinkedList<>();
-    
-    private ParsingHookRegistry() {
-        for (ParsingHook each : ServiceLoader.load(ParsingHook.class)) {
-            hooks.add(each);
-        }
-    }
-    
-    /**
-     * Get instance.
-     * 
-     * @return instance
-     */
-    public static ParsingHookRegistry getInstance() {
-        return INSTANCE;
-    }
+public interface ParsingHook {
     
     /**
      * Handle when parse started.
      *
      * @param sql SQL to be parsed
      */
-    public void start(final String sql) {
-        hooks.forEach(each -> each.start(sql));
-    }
+    void start(String sql);
     
     /**
      * Handle when parse finished success.
      *
      * @param sqlStatement sql statement
      */
-    public void finishSuccess(final SQLStatement sqlStatement) {
-        hooks.forEach(each -> each.finishSuccess(sqlStatement));
-    }
+    void finishSuccess(SQLStatement sqlStatement);
     
     /**
      * Handle when parse finished failure.
-     *
+     * 
      * @param cause failure cause
      */
-    public void finishFailure(final Exception cause) {
-        hooks.forEach(each -> each.finishFailure(cause));
-    }
+    void finishFailure(Exception cause);
 }
